Fixing the 'reconnect to hass bug'
At lease after 5 years \o/ !
This bug appeared because home assistant generates an error when trying
to auth on a connection already authenticated, so we were requeuing the
failed auth request indefinitely, failing again and again.
This happened because we might already have a auth connection in the
request tracker and we were trying to create a new one because the
server asked for it (with the type 'auth_required') or somehow, home
assistant sends 'auth_required' while we just authenticated. Not very
clear but, ignoring 'auth' requests if already authenticated fixes it.